Fix cast warnings

+++ lib/aliases_sysaix.h
@@ -65,7 +65,7 @@ aliases_hash (register const char *str, register size_t len)
       1040, 1040, 1040, 1040, 1040, 1040, 1040, 1040, 1040, 1040,
       1040, 1040, 1040, 1040, 1040, 1040, 1040, 1040
     };
-  register unsigned int hval = len;
+  register unsigned int hval = (unsigned int)len;
 
   switch (hval)
     {
